#2068 Move and remove files
