--- src/psyclone/tests/nemo/test_files/do_while.f90
@@ -1,56 +1,0 @@
-! -----------------------------------------------------------------------------
-! BSD 3-Clause License
-!
-! Copyright (c) 2019-2020, Science and Technology Facilities Council.
-! All rights reserved.
-!
-! Redistribution and use in source and binary forms, with or without
-! modification, are permitted provided that the following conditions are met:
-!
-! * Redistributions of source code must retain the above copyright notice, this
-!   list of conditions and the following disclaimer.
-!
-! * Redistributions in binary form must reproduce the above copyright notice,
-!   this list of conditions and the following disclaimer in the documentation
-!   and/or other materials provided with the distribution.
-!
-! * Neither the name of the copyright holder nor the names of its
-!   contributors may be used to endorse or promote products derived from
-!   this software without specific prior written permission.
-!
-! THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-! AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-! IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-! DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-! FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-! DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-! SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-! CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-! OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-! OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-! -----------------------------------------------------------------------------
-! Authors: A. R. Porter and N. Nobre, STFC Daresbury Lab
-
-program do_while_test
-  implicit none
-  integer :: ji, jj, jk
-  integer, parameter :: jpi=10, jpj=10, jpk=12
-  real, dimension(jpi,jpj,jpk) :: gdepw_0
-  real, dimension(jpi,jpj) :: rzisf_tbl
-  integer, dimension(jpi,jpj) :: mbkt
-  logical :: my_flag
-  
-  ! Test code with DO while loops
-  my_flag = .true.
-  do while(my_flag)
-     mbkt(:,:) = 0.0
-     my_flag = .false.
-  end do
-  
-  mbkt(:,:) = 0.0
-
-  jk = 2
-  DO WHILE ( jk .LE. mbkt(ji,jj) .AND. gdepw_0(ji,jj,jk) < rzisf_tbl(ji,jj) )
-     jk = jk + 1
-  END DO
-end program do_while_test
